Run SEIR model for Model 0

Runs the SEIR model using the Model 0 specification. https://drive.google.com/file/d/12Vppztxe6JdNUHEv9hMQZdmsLVs3elsJ/view?usp=sharing

SEIR_M0(S0, E0, I0, R0, beta.vector, num.days, influx, params)

Arguments

S0
Numeric. Initial number of susceptible individuals.

E0
Numeric. Initial number of exposed individuals.

I0
Numeric. Initial number of infected individuals.

R0
Numeric. Initial number of recovered individuals.

beta.vector
Vector, of beta values to use in the simulation on each day. Should have the same length as num.days.

num.days
Numeric. Number of days to run the simulation.

influx
List. This is hash consisting of a 'day' and an 'influx' number, meant to represent an influx of infections into the region

params
List. This should contain the parameters for running the SEIR, including: - Sigma - gamma.r: The rate at which non-hospitalized people recover. - gamma.h: The rate at which to-be hospitalized people in the general population get admitted to the hospital. - psi: The rate at which those that are hospitalized leave the hospital. - hosp.rate: percent hospitalized out of those infected - icu.rate: percent ICU admitted among those hospitalized - vent.rate: percent ventilated among those ICU admitted

Value

Dataframe, with susceptible counts, exposed counts, infected counts, hospitalization numbers, and recovery numbers numbers.

Details

The simulation ignores births and deaths.